refactor: split resource projection helpers
