+++ tests/Integration/test_monitor_resources_route.py
@@ -28,6 +28,15 @@ def test_monitor_resources_refresh_route_smoke():
     assert "refresh_status" in payload["summary"]
 
 
+def test_monitor_and_product_resource_routes_coexist_intentionally():
+    with TestClient(app) as client:
+        monitor_response = client.get("/api/monitor/resources")
+        product_response = client.get("/api/resources/overview")
+
+    assert monitor_response.status_code == 200
+    assert product_response.status_code == 200
+
+
 def test_monitor_health_route_smoke():
     with TestClient(app) as client:
         response = client.get("/api/monitor/health")
